+++ components/ModeChange.vue
@@ -8,7 +8,7 @@
 </template>
 
 <script lang="ts">
-import { useLayout } from '~/composables/layout'
+import { useLayout } from '~/hooks/useAppTheme'
 
 export default {
   setup() {
