+++ components/StripeLink.vue
@@ -24,7 +24,7 @@
 </template>
 
 <script lang="ts">
-import { useLayout } from '~/composables/layout'
+import { useLayout } from '~/hooks/useAppTheme'
 
 export default {
   setup() {
